lambda-wikipedia-search/requirements.txt
pydantic==1.10.1
python-dateutil==2.8.2
requests==2.28.1
